pyproject.toml
[project]
name = "investrak"
version = "0.1.0"
description = "A CLI financial portfolio and goal tracking application"
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "fpdf2>=2.7.6",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "freezegun>=1.0.0",  # For datetime testing
    "pytest-cov>=4.0.0"  # For coverage reporting
]

[project.scripts]
investrak = "investrak.cli.main:cli"
investrak-web = "investrak.web.run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3
